Move adapters/marc/ into adapters/transformers/marc/
Shared MARC parsing logic is only used by transformers (axiell, ebsco).
Colocate it with its consumers under transformers/marc/.
