+++ catalogue_graph/src/adapters/transformers/axiell_transformer.py
@@ -2,13 +2,13 @@ from datetime import datetime
 
 from pymarc.record import Record
 
-from adapters.marc.transformers.alternative_titles import extract_alternative_titles
-from adapters.marc.transformers.last_transaction_time import (
+from adapters.transformers.marc.alternative_titles import extract_alternative_titles
+from adapters.transformers.marc.last_transaction_time import (
     extract_last_transaction_time_to_datetime,
 )
-from adapters.marc.transformers.notes import extract_notes
-from adapters.marc.transformers.other_identifiers import extract_other_identifiers
-from adapters.marc.transformers.title import extract_title
+from adapters.transformers.marc.notes import extract_notes
+from adapters.transformers.marc.other_identifiers import extract_other_identifiers
+from adapters.transformers.marc.title import extract_title
 from adapters.transformers.marcxml_transformer import MarcXmlTransformer
 from adapters.utils.adapter_store import AdapterStore
 from ingestor.models.shared.invisible_reason import InvisibleReason
